Show profile in WindowBar and remove dashboard side menu
