--- lib/screens/dashboard/components/header.dart
@@ -1,126 +1,0 @@
-import 'package:admin/controllers/menu_app_controller.dart';
-import 'package:admin/responsive.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:provider/provider.dart';
-
-import '../../../constants.dart';
-import '../../../services/auth_service.dart';
-import '../../../features/users/users_page.dart';
-
-
-class Header extends StatelessWidget {
-  const Header({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        if (!Responsive.isDesktop(context))
-          IconButton(
-            icon: Icon(Icons.menu),
-            onPressed: context.read<MenuAppController>().controlMenu,
-          ),
-        if (!Responsive.isMobile(context))
-          Text(
-            "Dashboard",
-            style: Theme.of(context).textTheme.titleLarge,
-          ),
-        if (!Responsive.isMobile(context))
-          Spacer(flex: Responsive.isDesktop(context) ? 2 : 1),
-        Expanded(child: SearchField()),
-        ProfileCard()
-      ],
-    );
-  }
-}
-
-class ProfileCard extends ConsumerWidget {
-  const ProfileCard({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final auth = ref.watch(authServiceProvider);
-    if (auth == null) return const SizedBox.shrink();
-    return PopupMenuButton<String>(
-      onSelected: (value) {
-        if (value == 'logout') {
-          ref.read(authServiceProvider.notifier).logout();
-          Navigator.of(context).pop();
-        } else if (value == 'manage') {
-          Navigator.of(context)
-              .push(MaterialPageRoute(builder: (_) => const UsersPage()));
-        }
-      },
-      itemBuilder: (context) => [
-        if (auth.isAdmin)
-          const PopupMenuItem(
-              value: 'manage', child: Text('Gerenciar Acessos')),
-        const PopupMenuItem(value: 'logout', child: Text('Logout')),
-      ],
-      child: Container(
-        margin: EdgeInsets.only(left: defaultPadding),
-        padding: EdgeInsets.symmetric(
-          horizontal: defaultPadding,
-          vertical: defaultPadding / 2,
-        ),
-        decoration: BoxDecoration(
-          color: secondaryColor,
-          borderRadius: const BorderRadius.all(Radius.circular(10)),
-          border: Border.all(color: Colors.white10),
-        ),
-        child: Row(
-          children: [
-            Image.asset(
-              "assets/images/profile_pic.png",
-              height: 38,
-            ),
-            if (!Responsive.isMobile(context))
-              Padding(
-                padding:
-                    const EdgeInsets.symmetric(horizontal: defaultPadding / 2),
-                child: Text(auth.username),
-              ),
-            const Icon(Icons.keyboard_arrow_down),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class SearchField extends StatelessWidget {
-  const SearchField({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return TextField(
-      decoration: InputDecoration(
-        hintText: "Search",
-        fillColor: secondaryColor,
-        filled: true,
-        border: OutlineInputBorder(
-          borderSide: BorderSide.none,
-          borderRadius: const BorderRadius.all(Radius.circular(10)),
-        ),
-        suffixIcon: InkWell(
-          onTap: () {},
-          child: Container(
-            padding: EdgeInsets.all(defaultPadding * 0.75),
-            margin: EdgeInsets.symmetric(horizontal: defaultPadding / 2),
-            decoration: BoxDecoration(
-              color: primaryColor,
-              borderRadius: const BorderRadius.all(Radius.circular(10)),
-            ),
-            child: SvgPicture.asset("assets/icons/Search.svg"),
-          ),
-        ),
-      ),
-    );
-  }
-}
