+++ lib/screens/main/main_screen.dart
@@ -1,8 +1,8 @@
 import 'package:admin/controllers/menu_app_controller.dart';
-import 'package:admin/responsive.dart';
 import 'package:admin/screens/dashboard/dashboard_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:admin/widgets/window_bar.dart';
 
 import 'components/side_menu.dart';
 
@@ -11,26 +11,9 @@ class MainScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       key: context.read<MenuAppController>().scaffoldKey,
+      appBar: const WindowBar(title: 'Supervisão', showMenu: true),
       drawer: const SideMenu(current: SideMenuSection.dashboard),
-      body: SafeArea(
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            // We want this side menu only for large screen
-            if (Responsive.isDesktop(context))
-              Expanded(
-                // default flex = 1
-                // and it takes 1/6 part of the screen
-                child: SideMenu(current: SideMenuSection.dashboard),
-              ),
-            Expanded(
-              // It takes 5/6 part of the screen
-              flex: 5,
-              child: DashboardScreen(),
-            ),
-          ],
-        ),
-      ),
+      body: SafeArea(child: DashboardScreen()),
     );
   }
 }
